Merge Upstream Changes

+++ app/src/main/jni/vpnproxy-jni/pcap.h
@@ -41,8 +41,8 @@ typedef struct pcaprec_hdr_s {
     guint32_t orig_len;
 } __packed pcaprec_hdr_s;
 
-void write_pcap_hdr(int fd, const struct sockaddr *srv, size_t srv_size);
-void write_pcap_rec(int fd, const struct sockaddr *srv, size_t srv_size, const uint8_t *buffer, size_t length);
-size_t dump_pcap_rec(u_char *buffer, const u_char *pkt, size_t pkt_len);
+void write_pcap_hdr(int fd, const struct sockaddr *srv, int srv_size);
+void write_pcap_rec(int fd, const struct sockaddr *srv, int srv_size, const uint8_t *buffer, int length);
+size_t dump_pcap_rec(u_char *buffer, const u_char *pkt, int pkt_len);
 
 #endif // __MY_PCAP_H__
